SyncStatus model added, SyncStatusRepository added

+++ src/main/java/com/enesd/myshelfbackend/repository/elasticsearch/BookDocumentRepository.java
@@ -4,6 +4,6 @@ import com.enesd.myshelfbackend.model.documents.BookDocument;
 import com.enesd.myshelfbackend.repository.elasticsearch.custom.IBookDocumentRepositoryCustom;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface BookDocumentRepository extends ElasticsearchRepository<BookDocument, Integer>, IBookDocumentRepositoryCustom {
+public interface BookDocumentRepository extends ElasticsearchRepository<BookDocument, Long>, IBookDocumentRepositoryCustom {
 
 }
