+++ src/main/java/com/enesd/myshelfbackend/repository/elasticsearch/MediaContentDocumentRepository.java
@@ -4,5 +4,5 @@ import com.enesd.myshelfbackend.model.documents.MediaContentDocument;
 import com.enesd.myshelfbackend.repository.elasticsearch.custom.IMediaContentDocumentRepositoryCustom;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface MediaContentDocumentRepository extends ElasticsearchRepository<MediaContentDocument, Integer>, IMediaContentDocumentRepositoryCustom {
+public interface MediaContentDocumentRepository extends ElasticsearchRepository<MediaContentDocument, Long>, IMediaContentDocumentRepositoryCustom {
 }
